Switch NP flag format back to 32 bits (uint32_t)
This matches the corresponding change in the kernel.

+++ src/kernel_iface.c
@@ -88,7 +88,7 @@ int init_kernel_iface(void)
 	long page_size = sysconf(_SC_PAGESIZE);
 	void* mapped_at = NULL;
 
-	BUILD_BUG_ON(sizeof(union np_flag) != sizeof(uint64_t));
+	BUILD_BUG_ON(sizeof(union np_flag) != sizeof(uint32_t));
 
 	BUILD_BUG_ON(offsetof(struct control_page, sched.raw)
 		     != LITMUS_CP_OFFSET_SCHED);
